+++ providers.tf
@@ -1,6 +1,6 @@
 # Define the required Terraform providers and their versions
 terraform {
-  required_version = ">=1.6"
+  required_version = ">= 1.11"
 
   required_providers {
     google = {